app/requirements.txt
asgiref==3.4.1
Django==3.2.20
django-import-export==2.7.1
django-simple-history==3.0.0
matplotlib==3.7.3
gunicorn==21.2.0
psycopg2-binary==2.9.5
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.27.1
sqlparse==0.4.4
schedule==1.1.0
